requirements.txt
pyyaml==3.13
Mako==1.0.7